pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gemma4-ollama"
version = "2.0.0"
description = "Gemma4-Ollama python framework"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "Abel Garza Ramírez", email = "abel@ciberdata.net"},
]
dependencies = [
    "ollama",
    "requests",
    "python-dotenv",
    "gTTS",
    "customtkinter"
]

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
